Extract non selection parts of SelectionManager into ViewPicker
