+++ rviz_common/CMakeLists.txt
@@ -191,6 +191,7 @@ set(rviz_common_source_files
   src/rviz_common/interaction/selection_handler.cpp
   src/rviz_common/interaction/selection_manager.cpp
   src/rviz_common/interaction/selection_renderer.cpp
+  src/rviz_common/interaction/view_picker.cpp
   src/rviz_common/splash_screen.cpp
   # src/rviz_common/time_panel.cpp
   src/rviz_common/tool_manager.cpp
