+++ rviz_common/include/rviz_common/interaction/selection_manager.hpp
@@ -158,17 +158,6 @@ public:
   /// Change the size of the off-screen selection buffer texture.
   void setTextureSize(unsigned size) override;
 
-  /// Return true if the point at x, y in the viewport is showing an object, false otherwise.
-  /**
-   * If it is showing an object, result will be changed to contain the 3D point
-   * corresponding to it.
-   */
-  bool get3DPoint(
-    Ogre::Viewport * viewport,
-    int x,
-    int y,
-    Ogre::Vector3 & result_point) override;
-
   rviz_common::properties::PropertyTreeModel * getPropertyModel() override;
 
 private Q_SLOTS:
@@ -221,54 +210,8 @@ private:
   /// Unpacks a pixelbox into pixel_buffer_
   void unpackColors(const Ogre::PixelBox & box);
 
-  void setDepthTextureSize(unsigned width, unsigned height);
-
   void setUpSlots();
 
-  /// Gets the 3D points in a box around a point in a view port.
-  /**
-   * \param[in] viewport Rendering area clicked on.
-   * \param[in] x x coordinate of upper-left corner of box.
-   * \param[in] y y coordinate of upper-left corner of box.
-   * \param[in] width The width of the rendered box in pixels.
-   * \param[in] height The height of the rendered box in pixels.
-   * \param[in] skip_missing Whether to skip non-existing points or insert
-   *   NaNs for them
-   * \param[out] result_points The vector of output points.
-   *
-   * \returns True if any valid point is rendered in the box. NaN points count,
-   *   so if skip_missing is false, this will always return true if
-   *   width and height are > 0.
-   */
-  bool get3DPatch(
-    Ogre::Viewport * viewport,
-    int x,
-    int y,
-    unsigned width,
-    unsigned height,
-    bool skip_missing,
-    std::vector<Ogre::Vector3> & result_points);
-
-  /// Renders a depth image in a box around a point in a view port.
-  /**
-   * \param[in] viewport Rendering area clicked on.
-   * \param[in] x x coordinate of upper-left corner of box.
-   * \param[in] y y coordinate of upper-left corner of box.
-   * \param[in] width The width of the rendered box in pixels.
-   * \param[in] height The height of the rendered box in pixels.
-   * \param[out] depth_vector The vector of depth values.
-   *
-   * \returns True if rendering operation to render depth data to the depth
-   *   texture buffer succeeds.
-   *   Failure likely indicates a pretty serious problem.
-   */
-  bool getPatchDepthImage(
-    Ogre::Viewport * viewport,
-    int x,
-    int y,
-    unsigned width,
-    unsigned height,
-    std::vector<float> & depth_vector);
 
   DisplayContext * context_;
   HandlerManagerIface * handler_manager_;
@@ -294,11 +237,6 @@ private:
   static constexpr uint32_t kNumRenderTextures_ = 2;
   std::array<Ogre::TexturePtr, kNumRenderTextures_> render_textures_;
   std::array<Ogre::PixelBox, kNumRenderTextures_> pixel_boxes_;
-
-  // Graphics card -based depth finding of clicked points.
-  Ogre::TexturePtr depth_render_texture_;
-  uint32_t depth_texture_width_, depth_texture_height_;
-  Ogre::PixelBox depth_pixel_box_;
 
   Ogre::Rectangle2D * highlight_rectangle_;
   Ogre::SceneNode * highlight_node_;
