+++ rviz_common/src/rviz_common/view_controller.cpp
@@ -56,7 +56,7 @@
 #include "rviz_common/load_resource.hpp"
 #include "rviz_common/properties/bool_property.hpp"
 #include "rviz_common/properties/float_property.hpp"
-#include "rviz_common/interaction/selection_manager.hpp"
+#include "rviz_common/interaction/view_picker_iface.hpp"
 #include "rviz_common/render_panel.hpp"
 
 namespace rviz_common
@@ -229,10 +229,10 @@ void ViewController::handleKeyEvent(QKeyEvent * event, RenderPanel * panel)
 {
   Ogre::Viewport * viewport =
     rviz_rendering::RenderWindowOgreAdapter::getOgreViewport(panel->getRenderWindow());
-  if (event->key() == Qt::Key_F && viewport && context_->getSelectionManager()) {
+  if (event->key() == Qt::Key_F && viewport && context_->getViewPicker()) {
     QPoint mouse_rel_panel = panel->mapFromGlobal(QCursor::pos());
     Ogre::Vector3 point_rel_world;  // output of get3DPoint().
-    if (context_->getSelectionManager()->get3DPoint(viewport,
+    if (context_->getViewPicker()->get3DPoint(viewport,
       mouse_rel_panel.x(), mouse_rel_panel.y(),
       point_rel_world))
     {
