+++ rviz_common/test/mock_display_context.hpp
@@ -49,6 +49,7 @@ public:
   MOCK_CONST_METHOD0(getWindowManager, rviz_common::WindowManagerInterface * ());
   MOCK_CONST_METHOD0(getHandlerManager, rviz_common::interaction::HandlerManagerIface * ());
   MOCK_CONST_METHOD0(getSelectionManager, rviz_common::interaction::SelectionManagerIface * ());
+  MOCK_CONST_METHOD0(getViewPicker, rviz_common::interaction::ViewPickerIface * ());
   MOCK_CONST_METHOD0(getFrameManager, rviz_common::FrameManagerIface * ());
   MOCK_CONST_METHOD0(getFixedFrame, QString());
   MOCK_CONST_METHOD0(getFrameCount, uint64_t());
